improve HTTP message handling and fix method chaining
- Extract response handling logic into ResponseTrait
- Fix method chaining in HTTP requests to maintain fluent interface
- Correct argument order in Smith trait's dependencies injection
- Remove duplicate response interface implementations
- Implement createUri method in HTTPFactory
- Update authentication flow in tests
