+++ src/Lexicons/Traits/Endpoint.php
@@ -6,14 +6,13 @@ use Atproto\Contracts\HTTP\HTTPFactoryContract;
 use Atproto\Contracts\Lexicons\RequestContract;
 use Atproto\Contracts\Resources\ResponseContract;
 use Atproto\Factories\HTTPFactory;
-use Psr\Http\Message\MessageInterface;
-use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\StreamInterface;
-use Psr\Http\Message\UriInterface;
 
 trait Endpoint
 {
     use Lexicon;
+    use MessageTrait;
+    use RequestTrait;
+    use MessageAlias;
 
     private HTTPFactoryContract $factory;
     private RequestContract $request;
@@ -55,229 +54,12 @@ trait Endpoint
 
     private function initialize(): void
     {
-        $this->header('Accept', 'application/json')
-            ->header('Content-Type', 'application/json')
-            ->url('https://bsky.social')
+        $this->request = $this->url('https://bsky.social')
             ->path(sprintf('/xrpc/%s', $this->nsid()))
-            ->method('GET');
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getProtocolVersion(): string
-    {
-        return $this->request->getProtocolVersion();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function withProtocolVersion(string $version): MessageInterface
-    {
-        return $this->request = $this->request->withProtocolVersion($version);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getHeaders(): array
-    {
-        return $this->request->getHeaders();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function hasHeader(string $name): bool
-    {
-        return $this->request->hasHeader($name);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getHeader(string $name): array
-    {
-        return $this->request->getHeader($name);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getHeaderLine(string $name): string
-    {
-        return $this->request->getHeaderLine($name);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function withHeader(string $name, $value): MessageInterface
-    {
-        return $this->request = $this->request->withHeader($name, $value);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function withAddedHeader(string $name, $value): MessageInterface
-    {
-        return $this->request = $this->request->withAddedHeader($name, $value);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function withoutHeader(string $name): MessageInterface
-    {
-        return $this->request = $this->request->withoutHeader($name);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getBody(): StreamInterface
-    {
-        return $this->request->getBody();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function withBody(StreamInterface $body): MessageInterface
-    {
-        return $this->request = $this->request->withBody($body);
-    }
-
-    private function alias($name, ...$args)
-    {
-        if (($return = $this->request->$name(...$args)) instanceof RequestContract) {
-            $this->request = $return;
-
-            return $this;
-        }
-
-        return $return;
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function url($url = null)
-    {
-        return $this->alias(__FUNCTION__, $url);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function path(string $path = null)
-    {
-        return $this->alias(__FUNCTION__, $path);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function method(string $method = null)
-    {
-        return $this->alias(__FUNCTION__, $method);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function header(string $name, string $value = null)
-    {
-        return $this->alias(__FUNCTION__, $name, $value);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function parameter(string $name, $value = null)
-    {
-        return $this->alias(__FUNCTION__, $name, $value);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function queryParameter(string $name, $value = null)
-    {
-        return $this->alias(__FUNCTION__, $name, $value);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function headers($headers = null)
-    {
-        return $this->alias(__FUNCTION__, $headers);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function parameters($parameters = null)
-    {
-        return $this->alias(__FUNCTION__, $parameters);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function queryParameters($queryParameters = null)
-    {
-        return $this->alias(__FUNCTION__, $queryParameters);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getRequestTarget(): string
-    {
-        return $this->request->getRequestTarget();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function withRequestTarget(string $requestTarget): RequestContract
-    {
-        return $this->request = $this->request->withRequestTarget($requestTarget);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getMethod(): string
-    {
-        return $this->request->getMethod();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function withMethod(string $method): RequestContract
-    {
-        return $this->request = $this->request->withMethod($method);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getUri(): UriInterface
-    {
-        return $this->request->getUri();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function withUri(UriInterface $uri, bool $preserveHost = false): RequestInterface
-    {
-        return $this->request = $this->request->withUri($uri, $preserveHost);
+            ->method('GET')
+            ->headers([
+                'Accept' => 'application/json',
+                'Content-Type' => 'application/json'
+            ]);
     }
 }
